Committing the complete mess. I managed to trial&error on converting the php fantasy dictionaries into structs. Though some of the field types are mere blind guesses, so fingers crossed.

+++ Sources/Api/Record/MaxMindRecord.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 public struct MaxMindRecord {
-  // I don't even know any more what to say. Stupidest shit I saw in years.
-  let queriesRemaining: Int?
+  // assumed
+  let queriesRemaining: UInt16?
 }
